auth: implemented policies at controller level and route level with middleware

+++ app/Http/Controllers/IdeaController.php
@@ -6,6 +6,7 @@ use App\Models\Idea;
 use Illuminate\Http\Request;
 use App\Http\Requests\IdeaRequest;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
 
 class IdeaController extends Controller
 {
@@ -27,6 +28,7 @@ class IdeaController extends Controller
      */
     public function create()
     {
+//        Gate::authorize('create', Idea::class);
         return view('ideas.create');
     }
 
@@ -55,6 +57,13 @@ class IdeaController extends Controller
      */
     public function show(Idea $idea)
     {
+        Gate::authorize('update', $idea);
+
+        //other way to do it
+//        if(Auth::user()->cannot('update', $idea)) {
+//            return redirect('/ideas');
+//        }
+
         return view('ideas.show', [
             'idea' => $idea
         ]);
@@ -65,6 +74,8 @@ class IdeaController extends Controller
      */
     public function edit(Idea $idea)
     {
+//        Gate::authorize('update', $idea);
+
         return view('ideas.edit', [
             'idea' => $idea
         ]);
@@ -75,6 +86,7 @@ class IdeaController extends Controller
      */
     public function update(IdeaRequest $request, Idea $idea)
     {
+        Gate::authorize('update', $idea);
         $idea->update([
             'description' => request('description')
         ]);
@@ -86,6 +98,7 @@ class IdeaController extends Controller
      */
     public function destroy(Idea $idea)
     {
+        Gate::authorize('update', $idea);  // naka update lng kay same lng mn sila implementation
         $idea->delete(); //delete the specific idea
         return redirect('/ideas');
     }
